+++ src/linfo/__init__.py
@@ -8,7 +8,7 @@ Author: Roy Jensen <g04t@t3chg04t.wtf>
 ORCID: https://orcid.org/0009-0001-2601-8028
 """
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"
 __author__ = "Roy Jensen"
 __email__ = "g04t@t3chg04t.wtf"
 __orcid__ = "https://orcid.org/0009-0001-2601-8028"